docs + making domain input a float

+++ updaug/_models.py
@@ -54,7 +54,7 @@ def build_generator(num_domains, num_channels=3):
     
     
     inpt = tf.keras.layers.Input((None, None, num_channels))
-    domain_inpt = tf.keras.layers.Input((num_domains,), dtype=tf.int64)
+    domain_inpt = tf.keras.layers.Input((num_domains,), dtype=tf.float32)
     net = encoder(inpt)
     net = anorm(net, domain_inpt)
     net = decoder(net)
